+++ src/test/resources/expected-output/java/model/common/AddressV1Dto.java
@@ -2,13 +2,15 @@ package io.github.torand.test.model.common;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 @Schema(name = "AddressV1", description = "TBD")
 public record AddressV1Dto (
 
-    @Schema(description = "Street name and number", required = true)
+    @Schema(description = "Street name and number", required = true, minLength = 1, maxLength = 100)
     @NotBlank
+    @Size(min = 1, max = 100)
     String streetAddress,
 
     @Schema(description = "Zip code of postal area", required = true)